feat : #77 firebase folder structuring
