+++ src/components/ProfileIcon/index.tsx
@@ -9,8 +9,7 @@ import {
   isSelected,
   isNotSelected,
 } from './ProfileIcon.styled';
-import { db } from '../../firebase';
-import { collection, getDoc, getDocs } from 'firebase/firestore';
+import { getPostDocs } from '@/firebase/request';
 import { ListProps } from './List';
 
 interface ProfileIconProps {
@@ -20,26 +19,7 @@ interface ProfileIconProps {
 const ProfileIcon = ({ onClickToggleModal }: ProfileIconProps) => {
   const [modalOpened, setModalOpened] = useState(false);
   const [isLiFirst, setisLiFirst] = useState(true);
-  const [firstLiMockData, setFirstLiMockData] = useState<any>([
-    // {
-    //   title: '올레무스',
-    //   position: '서대문구',
-    //   type: '카페',
-    //   score: 4.4,
-    // },
-    // {
-    //   title: '바니프레소',
-    //   position: '강남구',
-    //   type: '카페',
-    //   score: 1.2,
-    // },
-    // {
-    //   title: '빽다방',
-    //   position: '강남구',
-    //   type: '카페',
-    //   score: 4.9,
-    // },
-  ]);
+  const [firstLiMockData, setFirstLiMockData] = useState<any>([]);
   const [secondLiMockData, setSecondLiMockData] = useState([
     {
       name: '담소순대국',
@@ -61,23 +41,11 @@ const ProfileIcon = ({ onClickToggleModal }: ProfileIconProps) => {
     },
   ]);
 
-  async function loadPostsData() {
-    const postsData = await getDocs(collection(db, 'posts'));
-    let posts: any = [];
-
-    postsData.forEach((post) => {
-      posts.push({ id: post.id, ...post.data() });
-    });
-
-    setFirstLiMockData(posts);
-  }
-
   useEffect(() => {
     // setSecondLiMockData([]); // 가고싶다 서브메뉴에 빈 데이터가 들어가는 경우의 테스트 코드
-    loadPostsData();
-    // getPostDocs().then((res) => {
-    //   setFirstLiMockData(res);
-    // });
+    getPostDocs().then((res) => {
+      setFirstLiMockData(res);
+    });
   }, []);
 
   const handleOpen = () => {
@@ -101,13 +69,6 @@ const ProfileIcon = ({ onClickToggleModal }: ProfileIconProps) => {
   const deleteBtnClick = () => {
     setFirstLiMockData([]);
   };
-
-  interface ListProps {
-    name: string;
-    address: string;
-    category: string;
-    score: number;
-  }
 
   return (
     <ModalContainer>
