+++ src/firebase/type.ts
@@ -1,0 +1,6 @@
+export interface Posts {
+  address: string;
+  category: string;
+  name: string;
+  score: number;
+}
