+++ src/index.tsx
@@ -4,7 +4,6 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from '@/store';
 import App from '@/App';
-import { auth } from './firebase';
 import { ThemeProvider } from '@emotion/react';
 import GlobalStyle from '@/styles/global';
 import theme from '@/styles/theme';
